+++ src/main/java/postman/bottler/user/auth/CustomUserDetailsService.java
@@ -5,7 +5,7 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Component;
 import postman.bottler.user.domain.User;
-import postman.bottler.user.service.UserRepository;
+import postman.bottler.user.applications.repository.UserRepository;
 
 @Component
 @RequiredArgsConstructor
