+++ src/main/java/postman/bottler/user/infra/ProfileImageRepositoryImpl.java
@@ -3,10 +3,9 @@ package postman.bottler.user.infra;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import postman.bottler.user.domain.ProfileImage;
-import postman.bottler.user.exception.EmailException;
 import postman.bottler.user.exception.ProfileImageException;
 import postman.bottler.user.infra.entity.ProfileImageEntity;
-import postman.bottler.user.service.ProfileImageRepository;
+import postman.bottler.user.applications.repository.ProfileImageRepository;
 
 @Repository
 @RequiredArgsConstructor
