+++ src/main/java/postman/bottler/user/presentation/UserController.java
@@ -1,4 +1,4 @@
-package postman.bottler.user.controller;
+package postman.bottler.user.presentation;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -31,7 +31,7 @@ import postman.bottler.user.exception.EmailException;
 import postman.bottler.user.exception.NicknameException;
 import postman.bottler.user.exception.PasswordException;
 import postman.bottler.user.exception.ProfileImageException;
-import postman.bottler.user.service.UserService;
+import postman.bottler.user.applications.UserService;
 
 @RestController
 @RequiredArgsConstructor
