+++ src/test/java/postman/bottler/user/applications/BanServiceTest.java
@@ -1,0 +1,8 @@
+package postman.bottler.user.applications;
+
+import org.junit.jupiter.api.DisplayName;
+
+@DisplayName("정지 서비스 테스트")
+class BanServiceTest {
+
+}
