+++ src/test/java/postman/bottler/user/applications/EmailServiceTest.java
@@ -1,0 +1,5 @@
+package postman.bottler.user.applications;
+
+class EmailServiceTest {
+
+}
