+++ src/test/java/postman/bottler/user/applications/KakaoServiceTest.java
@@ -1,0 +1,5 @@
+package postman.bottler.user.applications;
+
+class KakaoServiceTest {
+
+}
