+++ src/test/java/postman/bottler/user/applications/UserServiceTest.java
@@ -1,0 +1,9 @@
+package postman.bottler.user.applications;
+
+import org.junit.jupiter.api.DisplayName;
+
+@DisplayName("유저 서비스 테스트")
+class UserServiceTest {
+
+
+}
